add custom_tools.py
