+++ main.py
@@ -4,31 +4,23 @@ from src.handlers import create_handlers
 import asyncio
 from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
 handler = StreamingStdOutCallbackHandler()
-from langchain.agents import load_tools, Tool
 # Load environment variables
 load_dotenv()
 
-from langchain.tools import tool
-import subprocess
-@tool
-def disk_usage(query: str) -> str:
-    """useful for when you need to answer questions about the disk usage."""
-    output = subprocess.check_output(['df', '-h'], text=True)
-    output = "This is the output of `df -h`:\n" + output
-    return output
+# Load custom tools
+import src.custom_tools as custom_tools
 
-@tool
-def memory_usage(query: str) -> str:
-    """useful for when you need to answer questions about memory usage."""
-    output = subprocess.check_output(['free', '-h'], text=True)
-    output = "This is the output of `free -h`. Mem refers to RAM memory and Swap to swap memory:\n" + output
-    return output
+tools = [custom_tools.disk_usage, custom_tools.memory_usage,
+         custom_tools.asyncArxivQueryRun(max_workers=4),
+         custom_tools.asyncDuckDuckGoSearchRun(max_workers=4)]
 
-tools = [disk_usage, memory_usage]
-tools.extend(load_tools(['ddg-search', 'arxiv']))
+# You can load more tools using load_tools
+# from langchain.agents import load_tools
+# tools.extend(load_tools(['ddg-search', 'arxiv', 'requests_all']))
 
 # Create SlackBot instance
 bot = SlackBot(name='SlackBot', verbose=True,
+               max_tokens=500, model_type='openai',
                chunk_size=500, # Chunk size for splitter
                chunk_overlap=50, # Chunk overlap for splitter
                k_similarity=5, # Numbers of chunks to return in retriever
@@ -37,17 +29,16 @@ bot = SlackBot(name='SlackBot', verbose=True,
                )
 
 ## LLM configuration
-model_type = 'openai'
-if model_type == 'llama':
+if bot.model_type == 'llama':
     config = dict(gpu_layers=40, temperature=0.8, batch_size=1024, 
-                context_length=2048, threads=6, stream=True, max_new_tokens=500)
+                context_length=2048, threads=6, stream=True, max_new_tokens=bot.max_tokens)
 else:
-    config = dict(model_name="gpt-3.5-turbo", temperature=0.8, max_tokens=500)
+    config = dict(model_name="gpt-3.5-turbo-16k", temperature=0.8, max_tokens=bot.max_tokens)
 
 # Initialize LLM and embeddings
 bot.app.logger.info("Initializing LLM and embeddings...")
-bot.initialize_llm(model_type, max_tokens_threads=4000, config=config, callbacks=[handler])
-bot.initialize_embeddings(model_type)
+bot.initialize_llm(bot.model_type, max_tokens_threads=4000, config=config, callbacks=[handler])
+bot.initialize_embeddings(bot.model_type)
 
 # Create handlers for commands /ask, /modify_bot, /bot_info  and bot mentions
 create_handlers(bot)
@@ -65,8 +56,8 @@ async def start():
 if __name__ == "__main__":
     logger = bot.app.logger
     try:
-        logger.info('App started.')
         asyncio.run(start())
+        logger.info('App started.')
     except KeyboardInterrupt:
         logger.info('App stopped by user.')
     except Exception as e:
